catalog-toolkit/requirements.txt
squareup==44.0.1.20260122
requests>=2.32.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
